requirements.txt
# requirements.txt
langchain  # Core library for building language model workflows
python-dotenv  # For loading environment variables from .env files
google-generativeai  # Interact with Google's Generative AI models
langchain-google-genai  # LangChain integration for Google's Generative AI
flask  # Web framework for building APIs
flask-cors  # CORS support for Flask to handle cross-origin requests
langchain-community  # Additional document loaders and integrations for LangChain
transformers  # Hugging Face Transformers library for language models
sentence-transformers  # For pre-trained sentence embeddings
faiss-cpu  # FAISS library for vector similarity search
langchain-core
langchain-huggingface